+++ src/services/client/client-request-options.spec.ts
@@ -1,4 +1,5 @@
 import { ClientRequestOptions } from './client-request-options';
+import { mockAppConfig }        from '../context/app-config.mock';
 
 declare const expect, it, describe;
 
@@ -12,16 +13,17 @@ describe('ClientRequestOptions', () => {
         expect(opts.headers.get('Content-Type')).toBe('application/json');
         expect(opts.headers.get('X-Requested-With')).toBe('XMLHttpRequest');
         expect(opts.headers.get('X-Apigee-App-Id')).toBeFalsy();
+        expect(opts.headers.get('X-Apigee-App-Version')).toBeFalsy();
     });
 
     it('Building options with spa name', () => {
-        const spaName = 'randomSPAName';
-        const opts = new ClientRequestOptions(undefined, spaName);
+        const opts = new ClientRequestOptions(undefined, mockAppConfig);
 
         expect(opts.headers).toBeDefined();
         expect(opts.headers.get('Accept')).toBe('application/json');
         expect(opts.headers.get('Content-Type')).toBe('application/json');
         expect(opts.headers.get('X-Requested-With')).toBe('XMLHttpRequest');
-        expect(opts.headers.get('X-Apigee-App-Id')).toBe(spaName);
+        expect(opts.headers.get('X-Apigee-App-Id')).toBe(mockAppConfig.gtmAppName);
+        expect(opts.headers.get('X-Apigee-App-Version')).toBe(mockAppConfig.appVersion);
     });
 });